refactor(pdf): isolate preformatted panel classifiers
